Generation of Indexable files
Create a filter that allows us to take the entire original document collection, convert this into an array of ints, sort this and then use this array of ints to complete a binary search against the currently retrieved document collection to determine which documents are missing.

Once we determine these missing document IDs, then we create an XML of these documents in the same format as all of the retrieved ones.

We can then use this generated format to create the raw output of these missing documents so that they can also be indexed in the search engine.

The collection contains 678,706 found items, and 220,975 missing documents that have been created from  the original document collection.